# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ReviewGenie API - AI-powered review reply generator
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ReplyRequestTone = typeof ReplyRequestTone[keyof typeof ReplyRequestTone];


export const ReplyRequestTone = {
  Apologetic: 'Apologetic',
  Grateful: 'Grateful',
  Professional: 'Professional',
  Friendly: 'Friendly',
  Formal: 'Formal',
} as const;

/**
 * @nullable
 */
export type ReplyRequestPlatform = typeof ReplyRequestPlatform[keyof typeof ReplyRequestPlatform] | null;


export const ReplyRequestPlatform = {
  Google: 'Google',
  Facebook: 'Facebook',
} as const;

export interface ReplyRequest {
  /** @minLength 1 */
  reviewText: string;
  tone: ReplyRequestTone;
  /** @nullable */
  businessName?: string | null;
  /** @nullable */
  platform?: ReplyRequestPlatform;
  /** @nullable */
  starRating?: number | null;
}

export type ReplyResponseSentiment = typeof ReplyResponseSentiment[keyof typeof ReplyResponseSentiment];


export const ReplyResponseSentiment = {
  positive: 'positive',
  neutral: 'neutral',
  negative: 'negative',
} as const;

export interface ReplyResponse {
  reply: string;
  tone: string;
  sentiment: ReplyResponseSentiment;
  starRating: number;
}

export interface SentimentRequest {
  /** @minLength 1 */
  reviewText: string;
}

export type SentimentResponseSentiment = typeof SentimentResponseSentiment[keyof typeof SentimentResponseSentiment];


export const SentimentResponseSentiment = {
  positive: 'positive',
  neutral: 'neutral',
  negative: 'negative',
} as const;

export type SentimentResponseSuggestedTone = typeof SentimentResponseSuggestedTone[keyof typeof SentimentResponseSuggestedTone];


export const SentimentResponseSuggestedTone = {
  Apologetic: 'Apologetic',
  Grateful: 'Grateful',
  Professional: 'Professional',
  Friendly: 'Friendly',
  Formal: 'Formal',
} as const;

export interface SentimentResponse {
  sentiment: SentimentResponseSentiment;
  suggestedTone: SentimentResponseSuggestedTone;
  starRating: number;
}

export interface ErrorResponse {
  error: string;
}