# lib/api-zod/src/generated/types/replyRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ReviewGenie API - AI-powered review reply generator
 * OpenAPI spec version: 0.1.0
 */
import type { ReplyRequestPlatform } from './replyRequestPlatform';
import type { ReplyRequestTone } from './replyRequestTone';

export interface ReplyRequest {
  /** @minLength 1 */
  reviewText: string;
  tone: ReplyRequestTone;
  /** @nullable */
  businessName?: string | null;
  /** @nullable */
  platform?: ReplyRequestPlatform;
  /** @nullable */
  starRating?: number | null;
}