# lib/api-zod/src/generated/types/replyResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ReviewGenie API - AI-powered review reply generator
 * OpenAPI spec version: 0.1.0
 */
import type { ReplyResponseSentiment } from './replyResponseSentiment';

export interface ReplyResponse {
  reply: string;
  tone: string;
  sentiment: ReplyResponseSentiment;
  starRating: number;
}