# lib/api-zod/src/generated/types/sentimentRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ReviewGenie API - AI-powered review reply generator
 * OpenAPI spec version: 0.1.0
 */

export interface SentimentRequest {
  /** @minLength 1 */
  reviewText: string;
}