# lib/api-zod/src/generated/types/sentimentResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ReviewGenie API - AI-powered review reply generator
 * OpenAPI spec version: 0.1.0
 */
import type { SentimentResponseSentiment } from './sentimentResponseSentiment';
import type { SentimentResponseSuggestedTone } from './sentimentResponseSuggestedTone';

export interface SentimentResponse {
  sentiment: SentimentResponseSentiment;
  suggestedTone: SentimentResponseSuggestedTone;
  starRating: number;
}